requirements.txt
"""Requirements for Medical Relation Extraction project."""

# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# NLP libraries
spacy>=3.6.0
scispacy>=0.5.3
medspacy>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
click>=8.1.0

# Web framework for demo
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Optional: Advanced features
wandb>=0.15.0  # Experiment tracking
mlflow>=2.5.0  # Model management
opacus>=1.4.0  # Differential privacy
